(pure-move) migrate to_layout
