+++ tt_metal/impl/tensor/tensor_apis.cpp
@@ -295,6 +295,83 @@ std::vector<distributed::MeshCoordinate> enqueue_write_mesh_tensor(
 }  // namespace non_uniform_data_movement
 
 // ======================================================================================
+//                                  .to_layout()
+// ======================================================================================
+
+namespace {
+namespace CMAKE_UNIQUE_NAMESPACE {
+
+template <typename T>
+HostTensor to_layout_impl(const HostTensor& tensor, Layout target_layout) {
+    if (tensor.layout() == target_layout) {
+        return tensor;
+    }
+
+    auto source_layout = tensor.layout();
+    auto tile = tensor.tensor_spec().tile();
+    auto physical_shape = tensor.tensor_spec().physical_shape();
+    auto convert =
+        [tile, &physical_shape, source_layout, target_layout](const HostBuffer& input_host_buffer) -> std::vector<T> {
+        const auto input_data = input_host_buffer.view_as<T>();
+        switch (source_layout) {
+            case Layout::ROW_MAJOR:
+                TT_FATAL(target_layout == Layout::TILE, "Unsupported layout conversion");
+                return tensor_impl::convert_layout_row_major_to_tile(physical_shape, tile, input_data);
+            case Layout::TILE:
+                TT_FATAL(target_layout == Layout::ROW_MAJOR, "Unsupported layout conversion");
+                return tensor_impl::convert_layout_tile_to_row_major(physical_shape, tile, input_data);
+            case Layout::INVALID: TT_THROW("Invalid layout");
+        }
+        TT_THROW("Unreachable");
+    };
+
+    return HostTensor(
+        tensor.transform([&](const HostBuffer& buffer) { return HostBuffer(convert(buffer)); }),
+        TensorSpec(
+            tensor.logical_shape(),
+            TensorLayout::fromPaddedShape(
+                tensor.dtype(),
+                PageConfig(target_layout, tensor.tensor_spec().tile()),
+                MemoryConfig{},
+                tensor.logical_shape(),
+                tensor.padded_shape())),
+        tensor.tensor_topology());
+}
+
+template <typename T>
+HostTensor to_layout_bfloat_impl(const HostTensor& tensor, Layout target_layout) {
+    static_assert(
+        std::is_same_v<T, tensor_impl::bfloat8_b> || std::is_same_v<T, tensor_impl::bfloat4_b>, "Invalid type T");
+    // TODO: Flip to assert when we remove use cases in python and c++
+    if (tensor.layout() != target_layout or tensor.layout() != Layout::TILE) {
+        log_warning(
+            tt::LogAlways,
+            "Tensor layout must be Layout::TILE for bfloat8_b or bfloat4_b! Conversion from {} to {} was not executed!",
+            tensor.layout(),
+            target_layout);
+    }
+    return tensor;
+}
+
+template <>
+HostTensor to_layout_impl<tensor_impl::bfloat8_b>(const HostTensor& tensor, Layout target_layout) {
+    return to_layout_bfloat_impl<tensor_impl::bfloat8_b>(tensor, target_layout);
+}
+
+template <>
+HostTensor to_layout_impl<tensor_impl::bfloat4_b>(const HostTensor& tensor, Layout target_layout) {
+    return to_layout_bfloat_impl<tensor_impl::bfloat4_b>(tensor, target_layout);
+}
+
+}  // namespace CMAKE_UNIQUE_NAMESPACE
+}  // namespace
+
+HostTensor to_layout(const HostTensor& tensor, Layout target_layout) {
+    return tensor_impl::dispatch(
+        tensor.dtype(), [&]<typename T>() { return CMAKE_UNIQUE_NAMESPACE::to_layout_impl<T>(tensor, target_layout); });
+}
+
+// ======================================================================================
 //                                  .pad() and .unpad()
 // ======================================================================================
