+++ ttnn/api/ttnn/tensor/tensor_impl.hpp
@@ -23,12 +23,6 @@ namespace tt::tt_metal::tensor_impl {
 // ===============================================================================================================================================
 
 // ======================================================================================
-//                                  .to_layout()
-// ======================================================================================
-
-HostTensor to_layout(const HostTensor& tensor, Layout target_layout);
-
-// ======================================================================================
 //                                  .view()
 // ======================================================================================
